ambient-light/AmbientLightSensor_stop.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>AmbientLightSensor Test: stop()</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://www.w3.org/TR/ambient-light/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

let sensor, stop_return;

setup(function() {
  sensor = new AmbientLightSensor();
  sensor.start();
});

async_test(function(t) {
  sensor.onstatechange = t.step_func_done(function(event) {
    assert_equals(sensor.state, "idle");
  });
  sensor.onerror = t.step_func_done(function(event) {
    assert_unreached(event.error.name + ":" + event.error.message);
  });
  stop_return = sensor.stop();
}, "The sensor.state changes to 'idle' after sensor.stop()");

test(function() {
  assert_equals(sensor.reading, null);
}, "the sensor.reading is null after executing stop() method");

test(function() {
  assert_throws("InvalidStateError", function() { sensor.stop(); }, "stop() twice");
}, "throw an InvalidStateError exception when state is either idle or errored");

test(function() {
  assert_equals(stop_return, undefined);
}, "the sensor.stop() returns undefined");

</script>
